+++ packages/experimental-web-controls/src/Link/Link.helpers.ts
@@ -1,11 +1,10 @@
 'use strict';
 import * as React from 'react';
-import { ILinkRenderData, ILinkSettings, ILinkState } from './Link.types';
+import { ILinkRenderData, ILinkState, ILinkCustomizableProps, ILinkSlotProps } from './Link.types';
 import { IDivProps } from '../htmlTypes';
 import { useAsPressable, IWithOnStateChange, usePressableStateChange } from '../Pressable';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
-import { IAsResolved, renderSlot } from '@uifabricshared/foundation-composable';
-import { IThemeQueryInputs } from '@uifabricshared/foundation-compose';
+import { IUseComposeStyling } from '@uifabricshared/foundation-compose';
 
 export type retProp = { linkProps: IWithOnStateChange<IDivProps>; linkState: ILinkState; setState: (partial: ILinkState) => void };
 
@@ -29,55 +28,27 @@ export function useAsLink(props: IWithOnStateChange<IDivProps>): retProp {
   return { linkProps: newProps, linkState: state, setState: onSetState };
 }
 
-export function usePrepareState(data: ILinkRenderData): ILinkRenderData {
+export function useLinkPrepareProps(userProps: ILinkCustomizableProps, useStyling: IUseComposeStyling<ILinkSlotProps>): ILinkRenderData {
   // create the Link state/info once, re-renders happen with pressable state changes so this is storage
-  const { props, state } = useAsPressable(data.props);
-  const { linkProps, linkState } = useAsLink(data.props);
-  data.props = { ...props, ...linkProps };
-  data.state = {
-    ...state,
-    ...linkState,
-    URL: !!data.props.URL
+  const { props, state } = useAsPressable(userProps);
+  const { linkProps, linkState } = useAsLink(props);
+  linkProps.onClick = linkProps['internalOnClick'];
+
+  const URL = userProps.URL;
+  const resultState = { ...state, ...linkState, URL: !!URL };
+
+  // get the style slot props, using the state to resolve overrides
+  const styleSlotProps = useStyling(linkProps, resultState);
+
+  // setup the other props to get merged in
+  const transferProps = URL ? { root: linkProps } : { buttonAsRoot: linkProps };
+  const finalProps = {
+    root: { href: URL, URL: undefined, content: undefined },
+    content: { children: userProps.content }
   };
 
-  return data;
-}
+  // merge everything together
+  const slotProps = mergeSettings<ILinkSlotProps>(styleSlotProps, transferProps, finalProps);
 
-export function themeQueryInputs(name: string, renderData: ILinkRenderData): IThemeQueryInputs {
-  return { name, overrides: renderData.state };
-}
-
-export function finalizer(renderData: ILinkRenderData): ILinkRenderData {
-  const { props, slotProps } = renderData;
-  const final: ILinkSettings = { root: props };
-
-  if (props.content) {
-    final.content = { children: props.content };
-  }
-
-  // If we have a URL,
-  if (props.URL) {
-    // Only pass the URL prop to the href attribute if the Link is enabled
-    if (!props.disabled) {
-      final.root['href'] = props.URL;
-    }
-    delete props.URL;
-    delete final.buttonAsRoot;
-  } else {
-    final.buttonAsRoot = final.root;
-    delete final.root;
-  }
-
-  // Map internalOnClick over the original onClick; internalOnClick sets visited=true and calls the original onClick, if it exists
-  props.onClick = props['internalOnClick'];
-
-  renderData.slotProps = mergeSettings(slotProps, final);
-  return renderData;
-}
-
-export function view(result: IAsResolved<ILinkRenderData>, ...children: React.ReactNode[]): JSX.Element | null {
-  const slots = result.slots!;
-  const additionalChildren = children || [result.props.children];
-
-  return renderSlot(result.state.URL ? slots.root : slots.buttonAsRoot, renderSlot(slots.content), ...additionalChildren);
+  return { slotProps, state: resultState };
 }