+++ packages/experimental-web-controls/src/Stack/StackItem/StackItem.tsx
@@ -1,8 +1,8 @@
-import { IStackItemComponent } from './StackItem.types';
+import { IStackItemProps } from './StackItem.types';
 import { compose } from '@uifabricshared/foundation-compose';
 import { stackItemTokenProcessor } from './StackItem.tokens';
 
-export const StackItem = compose<IStackItemComponent>({
+export const StackItem = compose<IStackItemProps>({
   displayName: 'StackItem',
   settings: ['RNFStackItem'],
   slots: {