+++ packages/foundation-settings/src/Settings.types.ts
@@ -5,7 +5,6 @@ export interface ISlotProps<TProps extends object = object> {
 export type IPartialSlotProps<TSlotProps extends ISlotProps> = { [K in keyof TSlotProps]+?: Partial<TSlotProps[K]> };
 
 export type IComponentSettings<TSlotProps extends ISlotProps = ISlotProps> = IPartialSlotProps<TSlotProps> & {
-  _parent?: string | string[];
   _precedence?: string[];
   _overrides?: IComponentSettingsCollection<IComponentSettings<TSlotProps>>;
 };